chore: clarify json schema validator
